pyproject.toml
[project]
name = "query"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.11",
    "openai>=1.69.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.0",
]
